03-registro-tasa.test.ts with testing plan
